fix: allow consumers to access CssModule
Hey, 
This PR fixes a subtle breaking change happened in the last release.  
I had an integration that was injecting CssModule instances to the compilation, and it assumed this plugin will take care of them.
After the latest release it was impossible for me to get the CssModule class since it cannot be registered twice for serialization.
This PR allows others to get access to the CssModule class within same webpack running version.