Treat publicPath === ‘’ as meaning auto
We can’t do _nothing_ to publicPath, as the urls we output will not be correct if the css is output into a subfolder. This change means we treat publicPath ‘’ as meaning, my assets are available in the root of the webpack context, and please use relative paths to find them, and that’s what we do.