feat: support named exports with any characters
A proposed change to `css-loader` will allow named exports to be any
string, regardless of whether or not it is a valid identifier. The code
generation in this project assumes that all exports will be a valid
identifier, so the generation stage has been updated to take this into
account.